Working on a version of snake that can run in pygame
I have about 2 years of experience with python and about 3 months of experience with pygame. My code runs but it runs very slowly. I have looked over it with some people and we have both come to an agreement that there is a problem in my function or in my loop. However, due to my lack of experience with pygame I'm having trouble with identifying the issue.  I appreciate any kind of help with this.