Add files via upload
I've redone my code to the point where it's a lot more neater and easier to understand. However, when I run it it immediately stops and I've created print statements that show me what is causing it. The line of code that it causing the issue is the part meant for if the player were to hit themselves while they are playing. Condition 3 is the part that is causing the issue.